requirements.txt
openai==0.27.0
Flask==3.0.2
python-dotenv~=1.0.1